+++ ui/src/app/components/onboarding/Sections/WalletSection.tsx
@@ -1,0 +1,77 @@
+import { Button } from "@/app/components/buttons/Button";
+import { CompleteIcon } from "@/app/components/icons/Icons";
+import { displayAddress } from "@/app/lib/utils";
+import { getWalletConnectors } from "@/app/lib/connectors";
+import { useAccount, useConnect, useDisconnect } from "@starknet-react/core";
+
+interface WalletSectionProps {
+  step: number;
+}
+
+const WalletSection = ({ step }: WalletSectionProps) => {
+  const { address } = useAccount();
+  const { connectors, connect } = useConnect();
+  const { disconnect } = useDisconnect();
+  const walletConnectors = getWalletConnectors(connectors);
+  return (
+    <div className="relative z-1">
+      {step !== 1 && (
+        <>
+          <div className="absolute top-0 left-0 right-0 bottom-0 h-full w-full bg-black opacity-50 z-10" />
+          {step > 1 && (
+            <div className="absolute flex flex-col w-1/2 top-1/4 right-1/4 z-20 items-center text-xl text-center">
+              <p>Connected {displayAddress(address!)}</p>
+              <CompleteIcon />
+            </div>
+          )}
+        </>
+      )}
+      <div className="flex flex-col items-center justify-between sm:border sm:border-terminal-green p-5 text-center gap-10 z-1 h-[500px] sm:h-[425px] 2xl:h-[500px]">
+        <h4 className="m-0 uppercase text-3xl">Connect Starknet Wallet</h4>
+        <p className="sm:hidden 2xl:block text-xl">
+          In order to play LOOT SURVIVOR you are required to connect a Starknet
+          wallet.
+        </p>
+        <div className="hidden sm:flex flex-col">
+          {walletConnectors.map((connector, index) => (
+            <Button
+              disabled={address !== undefined}
+              onClick={() => {
+                disconnect();
+                connect({ connector });
+              }}
+              key={index}
+            >
+              {connector.id === "braavos" || connector.id === "argentX"
+                ? `Connect ${connector.id}`
+                : connector.id === "argentWebWallet"
+                ? "Login With Email"
+                : "Login with Cartridge Controller"}
+            </Button>
+          ))}
+        </div>
+        <div className="sm:hidden flex flex-col gap-2">
+          {walletConnectors.map((connector, index) => (
+            <Button
+              size={"lg"}
+              disabled={address !== undefined}
+              onClick={() => {
+                disconnect();
+                connect({ connector });
+              }}
+              key={index}
+            >
+              {connector.id === "braavos" || connector.id === "argentX"
+                ? `Connect ${connector.id}`
+                : connector.id === "argentWebWallet"
+                ? "Login With Email"
+                : "Login with Cartridge Controller"}
+            </Button>
+          ))}
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default WalletSection;
