+++ ui/src/app/lib/connectors.ts
@@ -1,7 +1,7 @@
 import { Connector } from "@starknet-react/core";
 import { InjectedConnector } from "starknetkit/injected";
-import { ArgentMobileConnector } from "starknetkit/argentMobile";
 import { WebWalletConnector } from "starknetkit/webwallet";
+import CartridgeConnector from "@cartridge/connector";
 
 export const checkArcadeConnector = (connector?: Connector) => {
   return typeof connector?.id === "string" && connector?.id.includes("0x");
@@ -44,9 +44,43 @@ export const argentWebWalletConnector = new WebWalletConnector({
   url: argentWebWalletUrl(),
 });
 
+const cartridgeConnector = new CartridgeConnector([
+  {
+    target: process.env.NEXT_PUBLIC_GAME_ADDRESS!,
+    method: "new_game",
+  },
+  {
+    target: process.env.NEXT_PUBLIC_GAME_ADDRESS!,
+    method: "explore",
+  },
+  {
+    target: process.env.NEXT_PUBLIC_GAME_ADDRESS!,
+    method: "attack",
+  },
+  {
+    target: process.env.NEXT_PUBLIC_GAME_ADDRESS!,
+    method: "flee",
+  },
+  {
+    target: process.env.NEXT_PUBLIC_GAME_ADDRESS!,
+    method: "equip",
+  },
+  {
+    target: process.env.NEXT_PUBLIC_GAME_ADDRESS!,
+    method: "drop",
+  },
+  {
+    target: process.env.NEXT_PUBLIC_GAME_ADDRESS!,
+    method: "upgrade",
+  },
+  {
+    target: process.env.NEXT_PUBLIC_LORDS_ADDRESS!,
+    method: "approve",
+  },
+]) as never as Connector;
+
 export const connectors = [
   new InjectedConnector({ options: { id: "braavos", name: "Braavos" } }),
   new InjectedConnector({ options: { id: "argentX", name: "Argent X" } }),
-  argentWebWalletConnector,
-  new ArgentMobileConnector(),
+  cartridgeConnector,
 ];
